requirements.txt
binance-connector
telethon
python-dotenv
asyncio
